Adds json linting to CodeMirror editor. Couldn't get it to work with karma so I used cdn script. This makes it possible to detect JSON errors while the config is being edited.
Makes re-auth button visible using inline-block display. This prevents it from moving to the next line.

Adds validation to environment variable UI_CONFIG. If an invalid config is provided, it is ignored and the error is logged if logging is enabled.

Provides clearer error message when Google provider is not configured for admin panel.

b/149712339

Change-Id: I74cdd723ddcbe0547752528834553f7aa92ec53f