Fixes token expiration errors in session_refresher iframe.
Currently the iframe will refresh every 45 minutes. But the Auth UI will not refresh the token if it is not yet expired and return a cached token. Since the ID token id 1 hour long, the token will expired after 15 minutes causing tokene expiration errors.

To fix this, we will always force token refresh. This is ok since IAP will only redirect to the auth ui if the token is expired. To prevent abuse, we will refresh within a threshold of 2 minutes from the time the token is issued.

Tested in https://ciap-prod-agent.appspot.com

b/154301670

Change-Id: I9ad1e0170db2d509e2d96d2558b7cee6f1d9ede2