Update the admin UI to use the IAP page URL as the authDomain.
This is to ensure that the <iap url>/admin page redirects to <iap url>/__/auth/handler
instead of firebaseapp.com/__/auth/handler, which will be a cross-origin call and run into
session storage issues.

This is a follow up of
https://team-review.git.corp.google.com/c/cicp-eng/cicp-iap-js/+/1676349
which addressed the issue for the main login page.

Verified with manual testing that the helloworld admin page in
testapp-iapupgrade project is recdirecting correctly:
https://helloworld-hsar65dtfq-uc.a.run.app/admin

b/303720834

Change-Id: I03daa66635cccceb9bdbfab445460798325f21b4